fix: stop the quantity floor from sticking on values like 0.29 in binary arithmetic

+++ src/Atria.Api/Controllers/Requests/AuthRequests.cs
@@ -152,7 +152,7 @@ public sealed record CreatePropertyRequest(
     string? Address,
     decimal TotalValue,
     decimal TokenPrice,
-    long TotalTokens,
+    decimal TotalTokens,
     string Currency,
     string? PropertyType = null,
     string? City = null,
@@ -295,7 +295,7 @@ public sealed record MarkReportFiledRequest(string FilingReference);
 /// <summary>POST /properties/{id}/annul-tokens body. Annuls unplaced capacity of an issue.</summary>
 /// <param name="TokenCount">How many unplaced shares to annul.</param>
 /// <param name="Reason">Why; required and journalled.</param>
-public sealed record AnnulTokensRequest(long TokenCount, string Reason);
+public sealed record AnnulTokensRequest(decimal TokenCount, string Reason);
 
 /// <summary>POST /properties/{id}/invalidate body. Declares an issue invalid (§73).</summary>
 /// <param name="Reason">The ground on which the issue is declared invalid; required and journalled.</param>
