+++ src/Atria.Application/Abstractions/Compliance/ITokenReader.cs
@@ -3,11 +3,11 @@ namespace Atria.Application.Abstractions;
 /// <summary>One share movement observed on chain.</summary>
 /// <param name="From">Sender; the zero address for a mint.</param>
 /// <param name="To">Recipient; the zero address for a burn.</param>
-/// <param name="Amount">Whole shares moved.</param>
+/// <param name="Amount">Shares moved, at the share granularity (decimals of a share).</param>
 /// <param name="BlockNumber">Block the transfer was mined in.</param>
 /// <param name="TransactionHash">Transaction that produced it.</param>
 public sealed record TokenTransfer(
-    string From, string To, long Amount, long BlockNumber, string TransactionHash);
+    string From, string To, decimal Amount, long BlockNumber, string TransactionHash);
 
 /// <summary>
 /// Reads an issue's token contract. Read-only by design: it holds no key, so nothing here can change
@@ -22,10 +22,10 @@ public sealed record TokenTransfer(
 public interface ITokenReader
 {
     /// <summary>Shares held by an address right now.</summary>
-    Task<long> GetBalanceAsync(string chainTag, string tokenContractAddress, string address, CancellationToken ct);
+    Task<decimal> GetBalanceAsync(string chainTag, string tokenContractAddress, string address, CancellationToken ct);
 
     /// <summary>Total shares in existence on the contract.</summary>
-    Task<long> GetTotalSupplyAsync(string chainTag, string tokenContractAddress, CancellationToken ct);
+    Task<decimal> GetTotalSupplyAsync(string chainTag, string tokenContractAddress, CancellationToken ct);
 
     /// <summary>The block the chain is currently at.</summary>
     Task<long> GetHeadBlockAsync(string chainTag, CancellationToken ct);
